+++ packages/gamut/Card/index.tsx
@@ -1,0 +1,3 @@
+export { default as CardShell } from './CardShell';
+export { default as CardBody } from './CardBody';
+export { default as CardFooter } from './CardFooter';