+++ packages/gamut/Icon/icons/ChevronDownIcon.tsx
@@ -1,0 +1,10 @@
+import React, { SVGProps } from 'react';
+
+export default function ChevronDownIcon(props: SVGProps<SVGSVGElement>) {
+  return (
+    <svg viewBox="0 0 24 24" {...props}>
+      <title>Chevron Down Icon</title>
+      <path d="M20.293 7.293l1.414 1.414L12 18.414 2.293 8.707l1.414-1.414L12 15.586z" />
+    </svg>
+  );
+}